fixup make run-on-kind
`make run-on-kind` was failing on the `kind-load-images` step because it
couldn't find the cluster. The create and delete steps use the
KIND_GPU_SIM_SCRIPT script defined in the kind-gpu-sim repo, where-as
load was calling kind directly. The script sets DOCKER_HOST and
KIND_EXPERIMENTAL_PROVIDER to allow podman to be used with kind. If
these variables are set in the local shell, everything works as expected.
But to make it easier on the user, the KIND_GPU_SIM_SCRIPT script was
extended to also support the load command.

If the user wants to access the cluster via kind (i.e. `kind get clusters`
or `kind get nodes`), DOCKER_HOST and KIND_EXPERIMENTAL_PROVIDER
still must be set in the local shell.

There was also a bug in the where the KIND_GPU_SIM_SCRIPT script was
being called. The script is expecting to be called like
`./kind-gpu-sim.sh delete --cluster-name=$(KIND_CLUSTER_NAME)`, but was
being called without the `=`, `./kind-gpu-sim.sh delete --cluster-name
$(KIND_CLUSTER_NAME)`. The script was just using default values instead
of taking input from the Makefile.

`make docker-build` was outputting the following WARNING, so I don't
think it was ever generating a usable image.

```
WARNING: No output specified with docker-container driver. Build result
will only remain in the build cache. To push result image into registry
use --push or to load image into docker use --load
```

Added `--load` to the docker build commands.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>