process configMap values
Process the LogLevels and NoGpu from the ConfigMap by passing the values
as Enviroment Variables from the deployment yamls. If the ConfigMap is
changed on the fly, these values will NOT be reapplied. May look into
that in the future.

The CSI Driver registration is also in the ConfigMap processing file. It
is a one and done and didn't need it's own controller. Reworked the
logic slightly so it's contained in it's own functions and less
intrusive to the ConfigMap processing.

Upstream CI is failing because the image being used in the unit tests,
"quay.io/gkm/vector-add-cache:rocm" was updated and a new SHA was
created. Longer term, need a way to lock the SHA, or the unit tests
don't really pull and extract an image. But that is for aother day, just
fix the hardcoded SHA for now.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>