revert makefile change
Small change in #50 changed the behavior of kustomize. For a command
like:

  `cd config/operator && $(KUSTOMIZE) edit set image quay.io/gkm/operator=${OPERATOR_IMG}`

The value on the left side of the `=`, in this case
`quay.io/gkm/operator` has to be what is in the kustomization.yaml file.
When a new image is being used (i.e. `${OPERATOR_IMG}` is different),
then what was coded was inserting a new section instead of overwriting
what was there.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>