handle MCV slow read better
AMD RoCm driver tends to be slow to detect GPUs and calls to MCV can
sometimes fail in this scenario. Have the Agent call MCV early in the
initialization code so the MCV cache is seeded with GPU data so this
delay is encountered once early in the cycle instead of when the
GKMCacheNode is being created. Also, handle errors better when MCV fails
or the GKMCacheNode CacheStatus is not present.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>